+++ src/commands/shared.ts
@@ -3,6 +3,10 @@ export const commandNames = {
     buildContract: 'ultra.buildContract',
     scaffoldContract: 'ultra.scaffoldContract',
     api: 'ultra.api',
+    destroyWallet: 'ultra.destroyWallet',
+    createWallet: 'ultra.createWallet',
+    showWalletPublicKeys: 'ultra.showWalletPublicKeys',
+    addKeyToWallet: 'ultra.addKeyToWallet',
 };
 
 export const installed: { [K in keyof typeof commandNames]: boolean } = {
@@ -10,4 +14,8 @@ export const installed: { [K in keyof typeof commandNames]: boolean } = {
     installHeaders: false,
     scaffoldContract: false,
     api: false,
+    destroyWallet: false,
+    createWallet: false,
+    showWalletPublicKeys: false,
+    addKeyToWallet: false,
 };