+++ src/utility/index.ts
@@ -4,4 +4,6 @@ export * as outputChannel from './outputChannel';
 export * as quickPick from './quickPick';
 export * as quickInput from './quickInput';
 export * as progress from './progressbar';
+export * as crypto from './crypto';
+export * as state from './state';
 export { sleep } from './sleep';